# taxmind-backend/docs/api-doc.yml
openapi: 3.0.0
info:
  title: TAXMIND-BACKEND
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Local server
...
paths:
...
  /admins/lookup:
    get:
      tags:
        - Admin
      summary: Lookup
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: '{{ACCESS_TOKEN}}'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins/profile:
    get:
      tags:
        - Admin
      summary: Profile
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: '{{ACCESS_TOKEN}}'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
  /admins:
    get:
      tags:
        - Admin
      summary: List Admin Users
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: '{{ACCESS_TOKEN}}'
        - name: page
          in: query
          schema:
            type: integer
          example: '1'
        - name: size
          in: query
          schema:
            type: integer
          example: '100'
        - name: keyword
          in: query
          schema:
            type: string
          example: 'jhon '
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - Admin
      summary: Register New User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: Nithen PV
                email: nithen.pv@ileafsolutions.com
                roleId: 992f8c7d-8e40-4f12-b3b9-59a48cf62c5e
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: '{{ACCESS_TOKEN}}'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /admins/dashboard:
    get:
      tags:
        - Admin
      summary: Dashboard
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: '{{ACCESS_TOKEN}}'
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...